requirements-full.txt
# Core dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
python-docx>=1.1.0

# LLM providers
anthropic>=0.18.0
openai>=1.12.0
google-genai>=0.2.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Utilities
rapidfuzz>=3.5.2
duckduckgo-search>=6.0.0
chardet>=5.2.0
python-dateutil>=2.8.2
oletools>=0.60.0

# Output
python-pptx>=0.6.23

# CLI
click>=8.1.7
colorama>=0.4.6
tabulate>=0.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Configuration
python-dotenv>=1.0.0

# Authentication
bcrypt>=4.1.0
argon2-cffi>=23.1.0
PyJWT>=2.8.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Supabase (optional - for Supabase Auth)
supabase>=2.0.0

# Narrative pipeline (optional)
pypdf>=3.0.0                # PDF fallback (lightweight)
unstructured[pdf]>=0.10.0    # PDF + layout-aware extraction
pyannote-audio>=3.0.0       # Speaker diarization (requires HUGGINGFACE_TOKEN)
